+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## Unreleased
 
+## 0.6.0 (2024-08-16)
+
 - Add new rules `enum-name-pascal-case` and `enum-value-snake-case`.
 - Add new rule `ban-unbounded-string-type`.
 - Add new rules `model-name-pascal-case` and `field-name-camel-case`.
+
+Thanks to @andyjy for these new rules!
 
 ## 0.5.0 (2024-05-01)